messageFuncs {shinyjs}	R Documentation

Show a message

Description

alert (and its alias info) shows a message to the user as a simple popup.

logjs writes a message to the JavaScript console. logjs is mainly used for debugging purposes as a way to non-intrusively print messages, but it is also visible to the user if they choose to inspect the console. You can also use the showLog function to print the JavaScript message directly to the R console.

Usage

alert(text)

info(text)

logjs(text)

Arguments

text
The message to show. Can be either simple text or an R object.

Note

shinyjs must be initialized with a call to useShinyjs() in the app's ui.

See Also

useShinyjs, runExample, showLog

Examples

if (interactive()) {
  library(shiny)
  shinyApp(
    ui = fluidPage(
      useShinyjs(),  # Set up shinyjs
      actionButton("btn", "Click me")
    ),
    server = function(input, output) {
      observeEvent(input$btn, {
        # Change the following line for more examples
        alert(paste0("The date is ", date()))
      })
    }
  )
}
## Not run: 
  # The shinyjs function call in the above app can be replaced by
  # any of the following examples to produce similar Shiny apps
  alert("Hello!")
  alert(text = R.Version())
  logjs(R.Version())

## End(Not run)

[Package shinyjs version 1.0 Index]