runExample {shinyjs}	R Documentation

Run shinyjs examples

Description

Launch a shinyjs example Shiny app that shows how to easily use shinyjs in an app.

Run without any arguments to see a list of available example apps. The "demo" example is also available online to experiment with.

Usage

runExample(example)

Arguments

example
The app to launch

Examples

## Only run this example in interactive R sessions
if (interactive()) {
  # List all available example apps
  runExample()

  runExample("sandbox")
  runExample("demo")
}

[Package shinyjs version 1.0 Index]