shinyjs-package {shinyjs}	R Documentation

shinyjs: Easily Improve the User Experience of Your Shiny Apps in Seconds

Description

Perform common useful JavaScript operations in Shiny apps that will greatly improve your apps without having to know any JavaScript. Examples include: hiding an element, disabling an input, resetting an input back to its original value, delaying code execution by a few seconds, and many more useful functions for both the end user and the developer. 'shinyjs' can also be used to easily call your own custom JavaScript functions from R.

Author(s)

Maintainer: Dean Attali daattali@gmail.com (ORCID)

See Also

Useful links:

https://deanattali.com/shinyjs/

Report bugs at https://github.com/daattali/shinyjs/issues

[Package shinyjs version 2.1.1 Index]