shinyjs {shinyjs}	R Documentation

shinyjs

Description

Easily improve the user experience of your Shiny apps in seconds

Details

shinyjs lets you perform common JavaScript operations that enhance the user experience in applications without having to know any JavaScript. Examples include: hiding an element, disabling an input, resetting an input back to its original value, delaying code execution by a few seconds, and many more useful functions. shinyjs also includes a colour picker widget, a colour picker RStudio addin, and can also be used to easily run your own custom JavaScript functions from R.

View the shinyjs website for more details and to see a demo.

[Package shinyjs version 1.0 Index]