showLog {shinyjs}	R Documentation

Print any JavaScript console.log messages in the R console

Description

When developing and debugging a Shiny that uses custom JavaScript code, it can be helpful to use console.log() messages in JavaScript. This function allows you to see these messages printed in the R console directly rather than having to open the JavaScript console in the browser to view the messages.

This function must be called in a Shiny app's server function, and you also need to pass the showLog=TRUE parameter to useShinyjs().

Usage

showLog()

Note

Due to an issue in shiny (see https://github.com/rstudio/shiny/issues/928), duplicated consecutive log messages will not get printed in R.

Log messages that cannot be serialized in JavaScript (such as many JavaScript Event objects that are cyclic) will not be printed in R.

See Also

logjs

Examples

if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useShinyjs(),
      textInput("text", "Type something")
    ),
    server = function(input, output) {
      showLog()
      logjs("App started")
      observe({
        logjs(paste("Length of text:", nchar(input$text)))
      })
    }
  )
}

[Package shinyjs version 1.0 Index]